Split up pagination into api.pagination and amo.pagination utils
