+++ src/olympia/amo/templatetags/jinja_helpers.py
@@ -110,7 +110,7 @@ def services_url(viewname, *args, **kwargs):
 
 @library.filter
 def paginator(pager):
-    return Paginator(pager).render()
+    return PaginationRenderer(pager).render()
 
 
 @library.filter
@@ -153,7 +153,7 @@ def sidebar(app):
     return categories, sorted(types, key=lambda x: x.name)
 
 
-class Paginator(object):
+class PaginationRenderer(object):
 
     def __init__(self, pager):
         self.pager = pager
